--- src/main/java/com/example/application/dto/GeoCodeResult.java
@@ -1,28 +1,0 @@
-package com.example.application.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class GeoCodeResult {
-
-    @JsonProperty("results")
-    GeoCode[] results;
-
-    @JsonProperty("generationtime_ms")
-    private String generationtimeMs;
-
-    public GeoCode[] getResults() {
-        return results;
-    }
-
-    public void setResults(GeoCode[] results) {
-        this.results = results;
-    }
-
-    public String getGenerationtimeMs() {
-        return generationtimeMs;
-    }
-
-    public void setGenerationtimeMs(String generationtimeMs) {
-        this.generationtimeMs = generationtimeMs;
-    }
-}
