+++ src/main/java/com/weather/application/Application.java
@@ -1,0 +1,31 @@
+package com.weather.application;
+
+import com.vaadin.flow.component.page.AppShellConfigurator;
+import com.vaadin.flow.server.PWA;
+import com.vaadin.flow.theme.Theme;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+/**
+ * The entry point of the Spring Boot application.
+ *
+ * Use the @PWA annotation make the application installable on phones, tablets
+ * and some desktop browsers.
+ *
+ */
+@SpringBootApplication
+@Theme(value = "flowcrmtutorial")
+@PWA(
+        name = "WeatherApp",
+        shortName = "WA",
+        offlinePath="offline.html",
+        offlineResources = { "./images/offline.png"}
+)
+public class Application extends SpringBootServletInitializer implements AppShellConfigurator {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}
