+++ src/main/java/com/weather/application/security/SecurityService.java
@@ -1,0 +1,16 @@
+package com.weather.application.security;
+
+import com.vaadin.flow.component.UI;
+import com.vaadin.flow.server.VaadinServletRequest;
+import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SecurityService  {
+
+    public void logout(){
+        UI.getCurrent().getPage().setLocation("/");
+        SecurityContextLogoutHandler logoutHandler = new SecurityContextLogoutHandler();
+        logoutHandler.logout(VaadinServletRequest.getCurrent().getHttpServletRequest(), null, null);
+    }
+}
