+++ src/main/java/com/weather/application/views/AbstractChartExample.java
@@ -1,0 +1,56 @@
+package com.weather.application.views;
+
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+
+import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.UI;
+import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.server.Command;
+
+/**
+ * Abstract class for all chart examples.
+ */
+public abstract class AbstractChartExample extends Div {
+    public AbstractChartExample() {
+        initDemo();
+    }
+
+    public abstract void initDemo();
+
+    protected void showNotification(String message) {
+        UI.getCurrent().getPage()
+                .executeJs("window.alert($0);", message);
+    }
+
+    /**
+     * Runs given task repeatedly until the reference component is attached
+     *
+     * @param component
+     * @param task
+     * @param interval
+     * @param initialPause
+     *            a timeout after tas is started
+     */
+    public static void runWhileAttached(Component component, Command task,
+                                        final int interval, final int initialPause) {
+        component.addAttachListener(event -> {
+            ScheduledExecutorService executor = Executors
+                    .newScheduledThreadPool(1);
+
+            component.getUI().ifPresent(ui -> ui.setPollInterval(interval));
+
+            final ScheduledFuture<?> scheduledFuture = executor
+                    .scheduleAtFixedRate(() -> {
+                        component.getUI().ifPresent(ui -> ui.access(task));
+                    }, initialPause, interval, TimeUnit.MILLISECONDS);
+
+            component.addDetachListener(detach -> {
+                scheduledFuture.cancel(true);
+                detach.getUI().setPollInterval(-1);
+            });
+        });
+    }
+}
