+++ src/main/java/com/weather/application/views/DashboardView.java
@@ -1,0 +1,115 @@
+package com.weather.application.views;
+
+
+import com.vaadin.flow.component.charts.Chart;
+import com.vaadin.flow.component.charts.model.AxisTitle;
+import com.vaadin.flow.component.charts.model.Configuration;
+import com.vaadin.flow.component.charts.model.DataSeries;
+import com.vaadin.flow.component.charts.model.Dimension;
+import com.vaadin.flow.component.charts.model.HorizontalAlign;
+import com.vaadin.flow.component.charts.model.Labels;
+import com.vaadin.flow.component.charts.model.LayoutDirection;
+import com.vaadin.flow.component.charts.model.Legend;
+import com.vaadin.flow.component.charts.model.PlotOptionsColumn;
+import com.vaadin.flow.component.charts.model.PlotOptionsSpline;
+import com.vaadin.flow.component.charts.model.Tooltip;
+import com.vaadin.flow.component.charts.model.VerticalAlign;
+import com.vaadin.flow.component.charts.model.XAxis;
+import com.vaadin.flow.component.charts.model.YAxis;
+import com.vaadin.flow.component.dependency.CssImport;
+import com.vaadin.flow.router.PageTitle;
+import com.vaadin.flow.router.Route;
+import jakarta.annotation.security.PermitAll;
+
+@Route(value = "dashboard", layout = MainLayout.class)
+@PageTitle("Dashboard | WeatherApp")
+@PermitAll
+@CssImport(value="./styles/MultipleAxes.css", themeFor = "vaadin-chart", include = "vaadin-chart-default-theme")
+public class DashboardView extends AbstractChartExample {
+    @Override
+    public void initDemo() {
+        Chart chart = new Chart();
+        Configuration conf = chart.getConfiguration();
+
+        conf.getChart().setZoomType(Dimension.XY);
+        conf.setTitle("Average Monthly Weather Data for Tokyo");
+        conf.setSubTitle("Source: WorldClimate.com");
+
+        XAxis x = new XAxis();
+        x.setCategories("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug",
+                "Sep", "Oct", "Nov", "Dec");
+        conf.addxAxis(x);
+
+        YAxis y1 = new YAxis();
+        y1.setShowEmpty(false);
+        y1.setTitle(new AxisTitle("Temperature"));
+        Labels labels = new Labels();
+        labels.setFormatter("return this.value +'°C'");
+        y1.setLabels(labels);
+        y1.setOpposite(true);
+        y1.setClassName("y1");
+        conf.addyAxis(y1);
+
+        YAxis y2 = new YAxis();
+        y2.setShowEmpty(false);
+        y2.setTitle(new AxisTitle("Rainfall"));
+        labels = new Labels();
+        labels.setFormatter("return this.value +' mm'");
+        y2.setLabels(labels);
+        y2.setClassName("y2");
+        conf.addyAxis(y2);
+
+        YAxis y3 = new YAxis();
+        y3.setShowEmpty(false);
+        y3.setTitle(new AxisTitle("Sea-Level Pressure"));
+        labels = new Labels();
+        labels.setFormatter("return this.value +' mb'");
+        y3.setLabels(labels);
+        y3.setOpposite(true);
+        y3.setClassName("y3");
+        conf.addyAxis(y3);
+
+        Tooltip tooltip = new Tooltip();
+        tooltip.setFormatter("function() { "
+                + "var unit = { 'Rainfall': 'mm', 'Temperature': '°C', 'Sea-Level Pressure': 'mb' }[this.series.name];"
+                + "return ''+ this.x +': '+ this.y +' '+ unit; }");
+        conf.setTooltip(tooltip);
+
+        Legend legend = new Legend();
+        legend.setLayout(LayoutDirection.VERTICAL);
+        legend.setAlign(HorizontalAlign.LEFT);
+        legend.setX(120);
+        legend.setVerticalAlign(VerticalAlign.TOP);
+        legend.setY(80);
+        legend.setFloating(true);
+        conf.setLegend(legend);
+
+        DataSeries series = new DataSeries();
+        PlotOptionsColumn plotOptionsColumn = new PlotOptionsColumn();
+        series.setPlotOptions(plotOptionsColumn);
+        series.setName("Rainfall");
+        series.setyAxis(1);
+        series.setData(49.9, 71.5, 106.4, 129.2, 144.0, 176.0, 135.6, 148.5,
+                216.4, 194.1, 95.6, 54.4);
+        conf.addSeries(series);
+
+        series = new DataSeries();
+        PlotOptionsSpline plotOptionsSpline = new PlotOptionsSpline();
+        series.setPlotOptions(plotOptionsSpline);
+        series.setName("Sea-Level Pressure");
+        series.setyAxis(2);
+        series.setData(1016, 1016, 1015.9, 1015.5, 1012.3, 1009.5, 1009.6,
+                1010.2, 1013.1, 1016.9, 1018.2, 1016.7);
+        conf.addSeries(series);
+
+        series = new DataSeries();
+        plotOptionsSpline = new PlotOptionsSpline();
+        series.setPlotOptions(plotOptionsSpline);
+        series.setName("Temperature");
+        series.setData(7.0, 6.9, 9.5, 14.5, 18.2, 21.5, 25.2, 26.5, 23.3, 18.3,
+                13.9, 9.6);
+        conf.addSeries(series);
+
+        add(chart);
+    }
+}
