+++ src/main/resources/META-INF/resources/offline.html
@@ -27,9 +27,9 @@
 <body>
 
 <div class="content">
-    <img src="./images/offline.png" alt="VaadinCRM is offline" class="offline-image"/>
+    <img src="./images/offline.png" alt="Weather App is offline" class="offline-image"/>
     <h1>Oh deer, you're offline</h1>
-    <p>Your internet connection is offline. Get back online to continue using Vaadin CRM.</p>
+    <p>Your internet connection is offline. Get back online to continue using Weather App.</p>
 </div>
 <script>
     window.addEventListener('online', () => window.location.reload());
