+++ src/main/resources/banner.txt
@@ -1,6 +1,6 @@
-  _____ _                  ____ ____  __  __   _____      _             _       _ 
- |  ___| | _____      __  / ___|  _ \|  \/  | |_   _|   _| |_ ___  _ __(_) __ _| |
- | |_  | |/ _ \ \ /\ / / | |   | |_) | |\/| |   | || | | | __/ _ \| '__| |/ _` | |
- |  _| | | (_) \ V  V /  | |___|  _ <| |  | |   | || |_| | || (_) | |  | | (_| | |
- |_|   |_|\___/ \_/\_/    \____|_| \_\_|  |_|   |_| \__,_|\__\___/|_|  |_|\__,_|_|
-                                                                                  
+ __        __         _   _                    _
+ \ \      / /__  __ _| |_| |__   ___ _ __     / \   _ __  _ __
+  \ \ /\ / / _ \/ _` | __| '_ \ / _ \ '__|   / _ \ | '_ \| '_ \
+   \ V  V /  __/ (_| | |_| | | |  __/ |     / ___ \| |_) | |_) |
+    \_/\_/ \___|\__,_|\__|_| |_|\___|_|    /_/   \_\ .__/| .__/
+                                                   |_|   |_|
